ciao-deploy: Pull more common code into utils.go for reuse
Restructure the existing code for reuse elsewhere. No functional change.
